Move CHERI auto demo to separate directory
Reorganises the automotive demo to place the CHERI-specific code
in its own CHERI directory, to mirror the organisation used for
non-CHERIoT legacy firmware, and to more clearly modularise and
separate the two implementations of the demo library.
