+++ examples/automotive/cheri/common.hh
